+++ src/app/logout/logout.component.ts
@@ -16,7 +16,6 @@ export class LogoutComponent implements OnInit {
     localStorage.removeItem('key');
     
     await this.userService.updateUser();
-    await this.userService.updateSavedUser();
 
     this.router.navigate(['/']);
   }