+++ src/environments/environment.prod.ts
@@ -2,6 +2,7 @@ export const environment = {
   production: true,
   endpoint: '/api',
   discordInvite: 'https://discord.gg/uDTgxyg',
-  version: 'v0.0.3a',
-  url: 'https://dbots.co'
+  version: 'v0.1.0b',
+  url: 'https://dbots.co',
+  guildId: '236608364333891585'
 };