+++ src/environments/environment.ts
@@ -6,11 +6,9 @@ export const environment = {
   production: false,
   endpoint: 'http://localhost:3000/api',
   discordInvite: 'https://discord.gg/uDTgxyg',
-  version: 'v0.0.3a',
-  url: 'https://dbots.co',
-  test: {
-    guildId: '685862664223850497'
-  }
+  version: 'v0.1.0b',
+  url: 'http://localhost:3000',
+  guildId: '531196495584821314'
 };
 
 /*